Introduce ProviderClients meta and extract provider configure helpers.
Return a typed client container from Configure instead of a bare PM/SM
client, so resource wrappers and a later Framework cutover share one
meta shape without changing user-facing provider behavior.
